requirements.txt
opencv-contrib-python
numpy
pandas
pillow
streamlit
fastapi
uvicorn
python-multipart
